pyproject.toml
[tool.poetry]
name = "apitest"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"
djoser = "^2.2.2"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.0"
python-dotenv = "^1.0.0"
phonenumbers = "^8.13.26"
django-phonenumber-field = "^7.2.0"
django-filter = "^23.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
